+++ R/dmt01.R
@@ -28,12 +28,12 @@ dmt01_main <- function(adam_db,
                          "ETHNIC",
                          "RACE"
                        ),
-                       lbl_overall = "All Patients",
+                       lbl_overall = "All {Patient_label}",
                        ...) {
   assert_valid_variable(adam_db$adsl, summaryvars)
   summaryvars_lbls <- var_labels_for(adam_db$adsl, summaryvars)
   assert_valid_variable(adam_db$adsl, c("USUBJID", arm_var), types = list(c("character", "factor")))
-
+  lbl_overall <- render_safe(lbl_overall)
   lyt <- dmt01_lyt(
     arm_var = arm_var,
     summaryvars = summaryvars,