+++ tests/testthat/test-aet05.R
@@ -4,7 +4,7 @@ test_that("aet05 can handle NA values", {
   proc_data$adaette$CNSR <- NA_real_
 
   res <- expect_silent(run(aet05, proc_data))
-  expect_snapshot(cat(formatters::export_as_txt(res, lpp = 100)))
+  expect_snapshot(cat(export_as_txt(res, lpp = 100)))
 })
 
 test_that("aet05 can handle some NA values", {
@@ -21,8 +21,8 @@ test_that("aet05 can handle some NA values", {
     )
 
   res1 <- expect_silent(run(aet05, proc_data))
-  expect_snapshot(cat(formatters::export_as_txt(res1, lpp = 100)))
+  expect_snapshot(cat(export_as_txt(res1, lpp = 100)))
 
   res2 <- expect_silent(run(aet05, proc_data, conf_level = 0.90, conf_type = "byar"))
-  expect_snapshot(cat(formatters::export_as_txt(res2, lpp = 100)))
+  expect_snapshot(cat(export_as_txt(res2, lpp = 100)))
 })